Utibu_Backend/requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
gunicorn==21.2.0
psycopg2==2.9.7
requests==2.31.0
flask_sqlalchemy
flask_migrate